+++ Sources/ACKLocalization/main.swift
@@ -1,0 +1,13 @@
+//
+//  main.swift
+//  
+//
+//  Created by Jakub Olejník on 11/12/2019.
+//
+
+import ACKLocalizationCore
+import Foundation
+
+let localization = ACKLocalization()
+
+localization.run()