+++ Sources/ACKLocalizationCore/Constants.swift
@@ -1,0 +1,14 @@
+//
+//  Constants.swift
+//  
+//
+//  Created by Jakub Olejník on 12/12/2019.
+//
+
+import Foundation
+
+/// Struct holding important constants used throughout the tool
+public struct Constants {
+    /// Prefix that defines that localization key holds plist item
+    public static let plistKeyPrefix = "plist"
+}