+++ Sources/ACKLocalizationCore/Model/SafeSubscript.swift
@@ -1,0 +1,15 @@
+//
+//  SafeSubscript.swift
+//  
+//
+//  Created by Jakub Olejník on 12/12/2019.
+//
+
+import Foundation
+
+internal extension RandomAccessCollection {
+    /// Returns element on `index` or `nil` if there isn't any
+    subscript(safe index: Index) -> Element? {
+        indices.contains(index) ? self[index] : nil
+    }
+}