+++ Tapestries/Sources/TapestryConfig/TapestryConfig.swift
@@ -1,0 +1,9 @@
+import PackageDescription
+
+let config = TapestryConfig(release: Release(actions: [.pre(.docsUpdate),
+                                                       .pre(.dependenciesCompatibility([.cocoapods, .carthage, .spm(.all)]))],
+                                             add: ["README.md",
+                                                   "ACKLocalization.podspec",
+                                                   "CHANGELOG.md"],
+                                             commitMessage: "Version \(Argument.version)",
+                                             push: false))